customizable scrollbar to sidebar config
