+++ data/src/config/pane.rs
@@ -1,11 +1,12 @@
 use serde::Deserialize;
 
+use crate::config::Scrollbar;
+
 #[derive(Debug, Clone, Deserialize, Default)]
 pub struct Pane {
     /// Default axis used when splitting a pane.
     #[serde(default)]
     pub split_axis: SplitAxis,
-    /// Scrollbar configuration
     #[serde(default)]
     pub scrollbar: Scrollbar,
 }
@@ -16,22 +17,4 @@ pub enum SplitAxis {
     #[default]
     Horizontal,
     Vertical,
-}
-
-#[derive(Debug, Copy, Clone, Deserialize, Default)]
-pub struct Scrollbar {
-    /// Width of the scrollbar.
-    #[serde(default = "default_scrollbar_width")]
-    pub width: u32,
-    /// Width of the scrollbar scroller.
-    #[serde(default = "default_scrollbar_scroller_width")]
-    pub scroller_width: u32,
-}
-
-fn default_scrollbar_width() -> u32 {
-    5
-}
-
-fn default_scrollbar_scroller_width() -> u32 {
-    5
 }
